requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
attrs==23.1.0
autopep8==2.0.4
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.4
defusedxml==0.7.1
flasgger==0.9.7.1
Flask==3.0.0
Flask-RESTful==0.3.10
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
pep8==1.7.1
pluggy==1.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.1
pytest==7.4.3
python-dotenv==1.0.1
python-simplexml==0.1.5
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.16.2
six==1.16.0
SQLAlchemy==2.0.27
tabulate==0.9.0
typing_extensions==4.10.0
urllib3==2.1.0
Werkzeug==3.0.1
xmltodict==0.13.0
